Few ui tweaks + found an issue
UI tweaks:
The title is now set correctly when selecting stations in a particular state (from the drawer).
The user can choose to see 'all' stations.

Issue:
The filter on the adapter causes the the adapter to cache the list of stations. This causes problems when trying to update it.