Fix for duplicate stations bug
This fixes a bug where multiple stations will show up when the station is the last in a region. The end of region </tr> caused the last station created to be added again.  Also added in an early exit condition if the observation reading and the observation data objects have not yet been initialised.

https://trello.com/c/DrQr4y10