fix station cache update bug
The station list cache was not being updated while the app stayed in memory. This is fixed by recording the time that the cache is taken, if the cache is becomes older than 15min then the next request will first return the stale cache, load a new cache in the background, then return the updated cache.

All tests pass.

https://trello.com/c/dOizJTbV